Move new journey page to index.tsx
